fix: tighten TUI menu reference contracts
